client-api: Move keys::claim_keys::v3 to a separate file
